pyproject.toml
[project]
name = "sql-database"
version = "0.1.0"
description = ""
authors = [
    {name = "Juan Veronez",email = "juan.monteirov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
